Backend/requirements.txt
asgiref==3.8.1
astroid==3.2.4
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coreschema==0.0.4
dill==0.3.8
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.7
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pylint==3.2.6
pylint-celery==0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.7.0
